feat(version-file): add version file support
This commit removes the use of the version-info lib and replaces it
with a python script and Zephyr's VERSION file system.

The python script version.py can be used to generate a valid VERSION
file based on the provided tag or based on the output of git describe.

The app firmware no linger uses the version-info lib and instead reads
the version from app_version.h, ncs_version.h, and other configs.