fix(infra): remove dockerfilelint hook from .pre-commit-config
The hook fails to install whenever a pre-commit command is run from a
clean state (delete .cache/pre-commit and try).

The issue seems to be some npm pack error due to a bad package in the
dependency chain.

Since dockerfilelint repo wasn't updated in the last 6 years is better
to completely remove this instead of trying to fix it.