docs: restructure docs
As we will soon transition from east to twister for the purposes of
release artefacts generations it is expected that some documentation
will have to be written for that.

As the current state of the template repo doesn't really have a nice
place to plug-in that documentation somewhere and since some time passed
since we were last updating the READMEs here, this commit does some
restructuring of the existing docs.

The general idea:
- Add doc/ folder. It should be used to contain all project specific
  documentation.
- Add doc/development folder. It should contain all documents that are
  somehow connected with the development aspect of the project.